+++ requirements.txt
@@ -6,3 +6,4 @@ ipython==5.1.0
 pylint==1.6.4
 PyMySQL==0.7.2
 PyYAML==3.11
+pyflakes==1.3.0
